project1.2/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
clip>=1.0.0
segment-anything>=1.0.0
opencv-python>=4.7.0
Pillow>=10.0.0
numpy>=1.24.0
gradio>=3.50.0
tqdm>=4.65.0 